Settings refactor
