+++ .env.example
@@ -1,1 +1,2 @@
 ENVIRONMENT=dev
+DEBUG=True
