+++ requirements/base.txt
@@ -1,3 +1,4 @@
 jupyter==1.0.0
 pydantic[dotenv]==2.4.2
+pydantic-settings==2.1.0
 PyYAML==6.0.1
